# Swagger/swagger.yaml
swagger: "2.0"
info:
  description: "Essa é a documentação das apis da CloudJob"
  version: "1.0.0"
  title: "Cloudjob"
host: "localhost:8080"
basePath: "/api"
...
schemes:
- "http"
paths:
  /candidatos:
    post:
      tags:
      - "Candidato"
      summary: "Adiciona um novo candidato"
      description: ""
      operationId: "addCandidato"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto de candidato que deve ser inserido"
        required: true
        schema:
          $ref: "#/definitions/Candidato"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
      - "Candidato"
      summary: "Acha todos os candidatos"
      description: "Retorna todos candidato"
      operationId: "getCandidatos"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Candidato"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Candidato not found"
      security:
      - api_key: []
...
  /vagas:
    post:
      tags:
      - "Vagas"
      summary: "Adiciona uma nova vaga"
      description: ""
      operationId: "addVaga"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto de vaga que deve ser inserido"
        required: true
        schema:
          $ref: "#/definitions/Vaga"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
      - "Vagas"
      summary: "Acha todas as vagas"
      description: "Retorna todas as vagas"
      operationId: "getVaga"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Vaga"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Vaga not found"
      security:
      - api_key: []
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Candidato:
    type: "object"
    properties:
      nome:
        type: "string"
      sobrenome:
        type: "string"
      dataDeNascimento:
        type: "string"
      email:
        type: "string"
      cpf:
        type: "string"
      pais:
        type: "string"
      cep:
        type: "string"
      descricao:
        type: "string"
      senha:
        type: "string"
      angular:
        type: "boolean"
        default: false 
      java:
        type: "boolean"
        default: false 
      groovy:
        type: "boolean"
        default: false 
      typescript:
        type: "boolean"
        default: false 
      python:
        type: "boolean"
        default: false 
...
  Vaga:
    type: "object"
    properties:
      nome:
        type: "string"
      descricao:
        type: "string"
      estado:
        type: "string"
      cidade: 
        type: "string"
      angular:
        type: "boolean"
        default: false 
      java:
        type: "boolean"
        default: false 
      groovy:
        type: "boolean"
        default: false 
      typescript:
        type: "boolean"
        default: false 
      python:
        type: "boolean"
        default: false 
...